Version 1.0.3
Added an experimental emacs-like feature. Vim already has a nice feature called 'incsearch' which is pretty useful. I have added a ^S command to the search prompt to take you to next search without having to press Enter and type n command. This is how you use it:

- open a file and make sure you have 'incsearch' on.
- start typing a search pattern, say /public
- without pressing Enter, Type ^S. You can see that you are taken to the next hit. Keep going.
- when you want to do a reverse search, just type ^X? and observe that you are now in ? prompt.
- you can start typing ^S again to go to the previous hits.
- when you want the cursor to move to a hit, just press enter as you would normally.
- if you want to cancel the search you can do so, but the original cursor position is restored only if you use ^U or Esc.

Also cleaned up the script a lot and fixed some misc. bugs. But before you start using the script, please be advised that it maps the ? / and : keys in normal mode. It tried to preserve your existing mapping for these keys (I have the / and ? mapped to turn on hlsearch and it continues to work even with the script), but I can't gaurantee that it will always work, so try the script and if you don't face any problems, you can start using it regularly. If you face any problems, please report it to me and I will try to solve it.

Please send in your comments.